+++ Runtime/Core/InternalsVisibleTo.cs
@@ -1,0 +1,2 @@
+using System.Runtime.CompilerServices;
+[assembly: InternalsVisibleTo("CaseyDeCoder.2Dx.Editor")]